+++ packages/app/lib/utils.dart
@@ -708,7 +708,7 @@ class Utils {
       try {
         return jsonDecode(config);
       } catch (e) {
-        // logger.d(e, error: e);
+        // logger.i(e, error: e);
       }
     }
     Storage.setString(StorageKeyString.defaultWebRTCServers,
@@ -799,6 +799,18 @@ class Utils {
         output: kReleaseMode
             ? LogFileOutputs(await Utils.createLogFile(directory.path))
             : null);
+
+    loggerNoLine = logger = Logger(
+        filter: kReleaseMode ? MyLogFilter() : null,
+        printer: PrettyPrinter(
+            dateTimeFormat: kDebugMode
+                ? DateTimeFormat.onlyTime
+                : DateTimeFormat.dateAndTime,
+            colors: false,
+            methodCount: kReleaseMode ? 1 : 0),
+        output: kReleaseMode
+            ? LogFileOutputs(await Utils.createLogFile(directory.path))
+            : null);
   }
 
   static bool isDomain(String str) {
@@ -878,20 +890,17 @@ class Utils {
   static Future<List<String>> waitRelayOnline(
       {List<String>? defaultRelays}) async {
     WebsocketService? ws = getGetxController<WebsocketService>();
-    int initAttempts = 0;
-    const maxInitAttempts = 5;
 
-    while (ws == null && initAttempts < maxInitAttempts) {
-      initAttempts++;
-      logger.d(
-          'Waiting for WebsocketService to initialize... ($initAttempts/$maxInitAttempts)');
+    for (int initAttempts = 1; initAttempts <= 5; initAttempts++) {
+      if (ws != null) break;
+      logger
+          .i('Waiting for WebsocketService to initialize... ($initAttempts/5)');
       await Future.delayed(const Duration(milliseconds: 300));
       ws = getGetxController<WebsocketService>();
     }
 
     if (ws == null) {
-      logger.e(
-          'Failed to initialize WebsocketService after $maxInitAttempts attempts');
+      logger.e('Failed to initialize WebsocketService');
       return [];
     }
     List<String> onlineRelays = ws.getOnlineSocketString();
@@ -899,22 +908,23 @@ class Utils {
     if (activeRelays.isEmpty) {
       activeRelays = ws.getActiveRelayString();
     }
-    int connectAttemptTimes = 0;
-    int connectMaxAttemptTimes = 5;
-
-    while (getIntersection(onlineRelays, activeRelays).isEmpty &&
-        connectAttemptTimes < connectMaxAttemptTimes) {
+    List<String> result = getIntersection(onlineRelays, activeRelays);
+    for (var checkRelayOnlineTimes = 0;
+        checkRelayOnlineTimes < 5;
+        checkRelayOnlineTimes++) {
+      if (result.isNotEmpty) {
+        return result;
+      }
       var debug = {
-        connectAttemptTimes: connectAttemptTimes,
+        checkRelayOnlineTimes: checkRelayOnlineTimes,
         'onlineRelays': onlineRelays,
         'activeRelays': activeRelays,
       };
-      logger.d('Waiting for relays to be available... $debug');
+      logger.i('Waiting for relays to be available... $debug');
       await Future.delayed(const Duration(seconds: 1));
-      connectAttemptTimes++;
       onlineRelays = ws.getOnlineSocketString();
     }
-    return getIntersection(onlineRelays, activeRelays);
+    return result;
   }
 
   static String _getDisplayName(String account, int nameLength) {
@@ -971,8 +981,8 @@ class Utils {
       EasyLoading.dismiss();
 
       return res;
-    } catch (e) {
-      EasyLoading.dismiss();
+    } catch (e, s) {
+      logger.e('Failed to create identity: $e', stackTrace: s);
       EasyLoading.showError("Failed to create identity: $e");
     }
     return null;
@@ -1018,22 +1028,21 @@ class Utils {
 
   // Catch and log errors to file
   static void logErrorToFile(String errorDetails) async {
-    if (kReleaseMode) {
-      try {
-        Directory appFolder = await getAppFolder();
-        final dateStr = DateFormat('yyyy-MM-dd').format(DateTime.now());
-        final String errorFilePath = '${appFolder.path}/errors/$dateStr.log';
-        final File file = File(errorFilePath);
+    if (kDebugMode) return;
+    try {
+      Directory appFolder = await getAppFolder();
+      final dateStr = DateFormat('yyyy-MM-dd').format(DateTime.now());
+      final String errorFilePath = '${appFolder.path}/errors/$dateStr.log';
+      final File file = File(errorFilePath);
 
-        if (!file.existsSync()) {
-          file.createSync(recursive: true);
-        }
-
-        await file.writeAsString('$errorDetails\n\n', mode: FileMode.append);
-        logger.e('Error written to $errorFilePath');
-      } catch (e) {
-        logger.e('Failed to write error to file: $e');
+      if (!file.existsSync()) {
+        file.createSync(recursive: true);
       }
+
+      await file.writeAsString('$errorDetails\n\n', mode: FileMode.append);
+      // logger.e('Error written to $errorFilePath');
+    } catch (e) {
+      logger.e('Failed to write error to file: $e');
     }
   }
 }